+++ pages/mypage/connect/index.tsx
@@ -1,7 +1,7 @@
 import type { CustomNextPage } from "next";
 import { FixedLayout } from "src/layout";
 
-import { Connect } from "../../../src/pages/mypage/connect/index";
+import { Connect } from "../../../src/components/page/mypage/connect/index";
 
 const ConnectPage: CustomNextPage = () => <Connect />;
