+++ pages/mypage/index.tsx
@@ -1,6 +1,6 @@
 import type { CustomNextPage } from "next";
+import { MyPage } from "src/components/page/mypage";
 import { FixedLayout } from "src/layout";
-import { MyPage } from "src/pages/mypage";
 
 const AboutPage: CustomNextPage = () => <MyPage />;
