+++ src/components/model/TaskWrap/index.tsx
@@ -4,8 +4,8 @@ import { useMemo } from "react";
 import { useCallback, useState } from "react";
 import { CgTrash } from "react-icons/cg";
 import { MdOutlineContentCopy } from "react-icons/md";
-import { RadioButton } from "src/component/RadioButton";
-import { TaskInput } from "src/component/taskInput";
+import { TaskInput } from "src/components/model/TaskInput";
+import { RadioButton } from "src/components/ui/RadioButton";
 import type { TodoType } from "src/lib/SupabaseClient";
 import { addTodo, deleteTodo, editIsComplete } from "src/lib/SupabaseClient";
 import type { BgColorProps, CaretColorProps, DayProps } from "src/type/type";
