+++ src/components/page/mypage/index.tsx
